pyproject.toml
[project]
name = "kalshi-temp-scanner"
version = "0.1.0"
description = "Read-only scanner for Kalshi daily temperature markets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "httpx>=0.27.0",
  "apscheduler>=3.10.4",
  "jinja2>=3.1.4",
  "pydantic-settings>=2.4.0",
  "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "respx>=0.21.1",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
